backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
ruff==0.6.2
pytest==8.3.2
pytest-django==4.8.0
nanoid==2.0.0
